examples/AWS_Strands/Product_Pete/requirements.txt
# AWS Strands Agent Dependencies
strands-agents
strands-agents-tools
python-dotenv
anthropic